backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
numpy==1.24.4
pillow==9.5.0
python-multipart==0.0.6
# If you trained the model with TensorFlow 2.20, change this to 2.20.0
# Installing TensorFlow on Windows can be easier via Conda. See README.md for details.
tensorflow==2.11.0
numpy
pillow
python-multipart
